Changed the timeout time to wait for all swap groups from 1.0 sec to 0.1 sec. Also created and using a different mutex to handle the signaling conditions. This must be tested on a larger IG cluster.
When the window is opened its added to a swap group (if swap groups are enabled/supported). Then when all windows in a cluster are opened and openGL is initiated the swapgroup is added to a swap barrier. This has proven to work better with more nvidia drivers then doing the whole process at once.